pyproject.toml
[project]
name = "cosmicwatch"
version = "0.1.0"
description = "AI-assisted space debris dashboard (Streamlit) with optional FastAPI + Next.js frontend"
requires-python = ">=3.11"
dependencies = [
    "ephem>=4.1.6",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "plotly>=6.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.25.0",
    "schedule>=1.2.0",
    "scikit-learn>=1.0.0",
    "scipy>=1.11.0",
    "seaborn>=0.11.0",
    "sgp4>=2.23",
    "sqlalchemy>=2.0.38",
    "tqdm>=4.62.0",
]
